+++ src/App/App.jsx
@@ -23,6 +23,7 @@ import {
 } from 'prop-types';
 
 import About from '../About/About';
+import { setDocumentationSections } from '../About/documentationSlice';
 import AppReloadDialog from '../AppReload/AppReloadDialog';
 import ErrorBoundary from '../ErrorBoundary/ErrorBoundary';
 import ErrorDialog from '../ErrorDialog/ErrorDialog';
@@ -87,6 +88,8 @@ const ConnectedApp = ({
     sidePanel,
     showLogByDefault = true,
     reportUsageData = false,
+    documentation,
+    children,
 }) => {
     const allPanes = useMemo(
         () => [...panes, { name: 'About', Main: About }].map(convertLegacy),
@@ -107,6 +110,10 @@ const ConnectedApp = ({
     useEffect(() => {
         dispatch(setPanes(allPanes));
     }, [dispatch, allPanes]);
+
+    useEffect(() => {
+        if (documentation) dispatch(setDocumentationSections(documentation));
+    }, [dispatch, documentation]);
 
     const SidePanelComponent = allPanes[currentPane].SidePanel;
     const currentSidePanel =
@@ -164,6 +171,7 @@ const ConnectedApp = ({
 
             <AppReloadDialog />
             <ErrorDialog />
+            {children}
         </div>
     );
 };
@@ -176,6 +184,12 @@ const PanePropType = exact({
     SidePanel: elementType,
 });
 
+const DocumentationPropType = exact({
+    title: string,
+    linkLabel: string,
+    link: string,
+});
+
 ConnectedApp.propTypes = {
     deviceSelect: node,
     panes: arrayOf(oneOfType([LegacyPanePropType, PanePropType]).isRequired)
@@ -183,6 +197,8 @@ ConnectedApp.propTypes = {
     sidePanel: node,
     showLogByDefault: bool,
     reportUsageData: bool,
+    documentation: arrayOf(DocumentationPropType),
+    children: node,
 };
 
 const noopReducer = (state = null) => state;