+++ src/Dropdown/Dropdown.tsx
@@ -6,28 +6,34 @@
 
 import React, { useEffect, useState } from 'react';
 import FormLabel from 'react-bootstrap/FormLabel';
-import { arrayOf, bool, func, number, string } from 'prop-types';
+import { arrayOf, bool, func, number, shape, string } from 'prop-types';
 
 import chevron from './chevron.svg';
 
-import './Dropdown.scss';
+import styles from './Dropdown.module.scss';
+
+interface DropdownItem {
+    label: string;
+    value: string;
+}
 
 interface DropdownProps {
     label?: string;
-    items: string[];
-    onSelect: (item: string, index: number) => void;
+    items: DropdownItem[];
+    onSelect: (item: DropdownItem) => void;
     disabled?: boolean;
     defaultIndex?: number;
+    truncateOptions?: boolean;
 }
 
 const Dropdown: React.FC<DropdownProps> = ({
     label,
     items,
-    disabled = false,
     onSelect,
-    defaultIndex,
+    disabled = false,
+    defaultIndex = 0,
 }) => {
-    const [selected, setSelected] = useState(items[defaultIndex ?? 0]);
+    const [selected, setSelected] = useState(items[defaultIndex]);
     const [isActive, setIsActive] = useState(false);
 
     useEffect(() => {
@@ -44,36 +50,38 @@ const Dropdown: React.FC<DropdownProps> = ({
 
     const onClick = () => setIsActive(!isActive);
 
+    const onClickItem = (item: DropdownItem) => {
+        setSelected(item);
+        onSelect(item);
+    };
+
     return (
-        <div className="dropdown-container">
-            {label && <FormLabel className="dropdown-label">{label}</FormLabel>}
+        <div className={styles.container}>
+            {label && <FormLabel className={styles.label}>{label}</FormLabel>}
             <button
                 type="button"
-                className={`dropdown-btn dropdown-btn-${
-                    isActive ? 'active' : 'inactive'
+                className={`${styles.btn} ${
+                    isActive ? styles.btnActive : styles.btnInactive
                 }`}
                 onClick={onClick}
                 disabled={disabled}
             >
-                <span>{selected}</span>
-                <img src={chevron} alt="" />
+                <span>{selected?.label}</span>
+                <img className={styles.image} src={chevron} alt="" />
             </button>
             <div
-                className={`dropdown-content dropdown-${
-                    isActive ? 'active' : 'inactive'
+                className={`${styles.content} ${
+                    isActive ? styles.itemsActive : styles.itemsInactive
                 }`}
             >
-                {items.map((item, index) => (
+                {items.map(item => (
                     <button
                         type="button"
-                        className="dropdown-item"
-                        key={item}
-                        onClick={() => {
-                            setSelected(item);
-                            onSelect(item, index);
-                        }}
+                        className={styles.item}
+                        key={item.value}
+                        onClick={() => onClickItem(item)}
                     >
-                        {item}
+                        {item.label}
                     </button>
                 ))}
             </div>
@@ -83,7 +91,12 @@ const Dropdown: React.FC<DropdownProps> = ({
 
 Dropdown.propTypes = {
     label: string,
-    items: arrayOf(string.isRequired).isRequired,
+    items: arrayOf(
+        shape({
+            label: string.isRequired,
+            value: string.isRequired,
+        }).isRequired
+    ).isRequired,
     onSelect: func.isRequired,
     disabled: bool,
     defaultIndex: number,