+++ src/index.js
@@ -17,12 +17,14 @@ export { default as ConfirmationDialog } from './Dialog/ConfirmationDialog';
 export { default as Spinner } from './Dialog/Spinner';
 export { default as Slider } from './Slider/Slider';
 export { default as Toggle } from './Toggle/Toggle';
+export { Alert } from './Alert/Alert';
 export { default as Main } from './Main/Main';
 export { default as Card } from './Card/Card';
 export { default as ErrorBoundary } from './ErrorBoundary/ErrorBoundary';
 export { default as StateSelector } from './StateSelector/StateSelector';
 export { default as Dropdown } from './Dropdown/Dropdown';
 export { default as StartStopButton } from './StartStopButton/StartStopButton';
+export { default as DocumentationSection } from './About/DocumentationSection';
 
 export { default as SidePanel } from './SidePanel/SidePanel';
 export { Group, CollapsibleGroup } from './SidePanel/Group';
@@ -45,6 +47,7 @@ export {
     getUserDataDir,
 } from './utils/appDirs';
 
+export { defaultInitPacket } from './Device/initPacket';
 export { openUrl } from './utils/open';
 export { default as systemReport } from './utils/systemReport';
 export { default as usageData } from './utils/usageData';
@@ -57,5 +60,12 @@ export { currentPane, setCurrentPane } from './App/appLayout';
 export { getAppSpecificStore as getPersistentStore } from './utils/persistentStore';
 
 export { deviceInfo } from './Device/deviceInfo/deviceInfo';
-export { getDeviceLibContext } from './Device/deviceLister';
+export {
+    getDeviceLibContext,
+    stopWatchingDevices,
+    startWatchingDevices,
+    waitForDevice,
+} from './Device/deviceLister';
 export { prepareDevice } from './Device/deviceSetup';
+export { default as sdfuOperations } from './Device/sdfuOperations';
+export { HashType, FwType } from './Device/initPacket';